requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
python-decouple==3.8
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0
